PhotoController has been added.
