+++ Tinder/Controllers/User Details/PhotoController.swift
@@ -1,0 +1,33 @@
+//
+//  PhotoController.swift
+//  Tinder
+//
+//  Created by Turan Çabuk on 16.02.2024.
+//
+
+import UIKit
+
+class PhotoController: UIViewController {
+    
+    let imageView = UIImageView(image: #imageLiteral(resourceName: "questionMark"))
+    
+    init(imageUrl: String) {
+        if let url = URL(string: imageUrl) {
+            imageView.sd_setImage(with: url)
+        }
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.addSubview(imageView)
+        imageView.fillSuperview()
+        imageView.contentMode = .scaleAspectFill
+        imageView.clipsToBounds = true
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
