+++ Tinder/Controllers/User Details/SwipingPhotosController.swift
@@ -68,26 +68,3 @@ class SwipingPhotosController: UIPageViewController, UIPageViewControllerDataSou
         }
     }
 }
-
-class PhotoController: UIViewController {
-    
-    let imageView = UIImageView(image: #imageLiteral(resourceName: "top_left_profile"))
-    
-    init(imageUrl: String) {
-        if let url = URL(string: imageUrl) {
-            imageView.sd_setImage(with: url)
-        }
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.addSubview(imageView)
-        imageView.fillSuperview()
-        imageView.contentMode = .scaleAspectFill
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
